+++ test/rocprim/test_device_search_n.cpp
@@ -404,7 +404,7 @@ TYPED_TEST(RocprimDeviceSearchNTests, SmallCount)
     constexpr bool debug_synchronous = TestFixture::debug_synchronous;
     op_type        op{};
 
-    for(size_t seed_index = 0; seed_index < random_seeds_count + seed_size; seed_index++)
+    for(size_t seed_index = 0; seed_index < number_of_runs; seed_index++)
     {
         unsigned int seed_value
             = seed_index < random_seeds_count ? rand() : seeds[seed_index - random_seeds_count];